requirements.txt
Flask
Flask-RESTful
Flask-JWT
flask-cors
mysql-connector-python
Flask-SQLAlchemy
pymysql
gunicorn

